+++ collective/recipe/solrinstance/__init__.py
@@ -64,7 +64,8 @@ NOT_ALLOWED_ATTR = set(["index", "filter", "unique-key", "max-num-results",
                         "additional-solrconfig",
                         "additional-solrconfig-query",
                         "autoCommitMaxDocs", "autoCommitMaxTime", "updateLog",
-                        "requestParsers-multipartUploadLimitInKB"])
+                        "requestParsers-multipartUploadLimitInKB",
+                        "requestParsers-enableRemoteStreaming"])
 
 
 class SolrBase(object):
@@ -197,6 +198,9 @@ class SolrBase(object):
             'additional-schema-config', '').strip()
         options['requestParsers-multipartUploadLimitInKB'] = options_orig.get(
             'requestParsers-multipartUploadLimitInKB', '102400').strip()
+        options['requestParsers-enableRemoteStreaming'] = options_orig.get(
+            'requestParsers-enableRemoteStreaming', 'false'
+        )
         options['extraFieldTypes'] = options_orig.get('extra-field-types', '')
 
         options['mergeFactor'] = options_orig.get('mergeFactor', '10')
@@ -601,6 +605,8 @@ class SolrSingleRecipe(SolrBase):
             maxWarmingSearchers=self.solropts.get('maxWarmingSearchers', '4'),
             requestParsers_multipartUploadLimitInKB=self.solropts[
                 'requestParsers-multipartUploadLimitInKB'],
+            requestParsers_enableRemoteStreaming=self.solropts[
+                'requestParsers-enableRemoteStreaming'],
             autoCommit=self.parseAutoCommit(self.solropts),
             updateLog=self.solropts.get('updateLog', 'false'),
             mergeFactor=self.solropts['mergeFactor'],
@@ -770,6 +776,8 @@ class MultiCoreRecipe(SolrBase):
                                                      '4'),
                 requestParsers_multipartUploadLimitInKB=options_core[
                     'requestParsers-multipartUploadLimitInKB'],
+                requestParsers_enableRemoteStreaming=options_core[
+                    'requestParsers-enableRemoteStreaming'],
                 autoCommit=self.parseAutoCommit(options_core),
                 updateLog=options_core['updateLog'],
                 mergeFactor=options_core['mergeFactor'],