+++ CHANGES.rst
@@ -5,7 +5,8 @@ Change History
 5.3.3 (unreleased)
 ==================
 
-- Nothing changed yet.
+- Update request handler in ``solrconfig.xml`` to avoid deprecation warnings
+  for SOLR 4.  [mgrbyte]
 
 
 5.3.2 (2014-08-27)
@@ -24,6 +25,10 @@ Change History
   generated.
   [davidjb]
 - Fix solrconfig.xml for indexing binary files [giacomos]
+- Added ``requestParsers-enableRemoteStreaming`` to allow configuring remote
+  streaming [wengole]
+- Changed default for remote streaming in solr 4 templates back to false (the
+  solr default) [wengole]
 
 
 5.3.1 (2013-12-17)